[new feat] add DualUMM
* [new fea] DualUMMv1 & DualUMMv2
* [bugfix] compatible with the new version of get_filesystem
* [reformat] reformat hdfs_model_checkpoint.py
* [reformat] reformat hdfs_model_checkpoint.py
* [remove] remove unused file: attentions.py
* [bugfix] force text_ids to be torch.LongTensor
* [cleaning] remove some unused files in voc_modules
* [new feat] add config of 160bins melgan
* [remove] remove WIP DualUMMv3
* [rename] rename var in training_step_disc
* [rename] rename var in training_step

https://code.byted.org/seed/samantha/merge_requests/712

+++ recipes/umm/modules/lit_module.py
@@ -1262,7 +1262,7 @@ class Stage2MSS(Stage2):
     @torch.cuda.amp.autocast(enabled=False)
     def prepare_feature(self, batch):
         # Prepare tokens
-        input_dict = {"text_ids": batch["token"]}
+        input_dict = {"text_ids": batch["token"].long()}
         # Prepare MSS audio tracks
         _audio_dict = {k: batch[k] for k in ["audio", "audio_vocal", "audio_inst"]}
         _audio_dict = {k: v.squeeze(dim=1).float() for k, v in _audio_dict.items()}
